Different payload types due to partial fix
An improvement was made to python-engineio master.

Now the `decode` function doesn't have `errors='ignore'` so we can't
include any bad utf-8 in later versions. You can choose which payload to
send depending on your target.

Another method of DoS was discovered, involving a single giant packet
with an integral payload. This causes recursion in python-socketio
packet decoding which locks up the server.